// src/bot/types/botgram.d.ts
  interface NetworkError extends Error {
    req: RequestInfo;
  }

  interface TelegramError {
    req: RequestInfo;
  }

  type RequestError = TelegramError | NetworkError;
// ...
  interface ReplyQueue {
    text(text: string, mode?: ParseMode): this;

    html(text: string, ...args: string[]): this;

    markdown(text: string): this;

    photo(file: model.InputFile, caption?: string, captionMode?: ParseMode): this;

    audio(file: model.InputFile, duration?: number, performer?: string, title?: string, caption?: string, captionMode?: ParseMode): this;

    document(file: model.InputFile, caption?: string, captionMode?: ParseMode): this;

    sticker(file: model.InputFile): this;

    video(file: model.InputFile, duration?: number, width?: number, height?: number, caption?: string, captionMode?: ParseMode, streaming?: boolean): this;

    videoNote(file: model.InputFile, duration?: number, length?: number): this;

    voice(file: model.InputFile, duration?: number, caption?: string, captionMode?: ParseMode): this;

    location(latitude: number, longitude: number): this;

    venue(latitude: number, longitude: number, title: string, address: string, foursquareId?: string): this;

    contact(phone: string, firstname: string, lastname?: string): this;

    game(gameShortName: string): this;

    command(command: string, ...args: string[]): this;

    command(explicit: boolean, command: string, ...args: string[]): this;

    forward(id: number, chat: model.ChatLike): this;

    forward(msg: Message): this;

    message(msg: Message, reforward?: boolean): this;

    silent(silent?: boolean): this;

    disablePreview(disable?: boolean): this;

    selective(selective?: boolean): this;

    forceReply(force?: boolean): this;

    keyboard(keys: model.KeyboardRow[] | false | null, resize?: boolean, oneTime?: boolean): this;

    reply(msg?: model.MessageLike | false): this;

    inlineKeyboard(keys: model.InlineKeyboardButton[][]): this;

    action(action?: ChatAction): this;

    editText(msg: model.MessageLike | string, text: string, mode?: ParseMode): this;

    editHTML(msg: model.MessageLike | string, text: string, ...args: string[]): this;

    editMarkdown(msg: model.MessageLike | string, text: string): this;

    editReplyMarkup(msg: model.MessageLike | string): this;

    editCaption(msg: model.MessageLike | string, caption: string): this;

    deleteMessage(msg: model.MessageLike): this;

    then(): Promise<any>;

    then(callback: (err: RequestError | null, result?: any) => any): this;

    to(chat: model.ChatLike): ReplyQueue;

    // not official
    sendGeneric(arg0: string, arg1: { text: string; parse_mode?: string; disable_web_page_preview?: boolean; }): void;
  }

  type ParseMode = null | "Markdown" | "HTML";
  type ChatAction =
    "typing" |
    "find_location" |
    "upload_photo" |
    "record_video" |
    "upload_video" |
    "record_audio" |
    "upload_audio" |
    "record_video_note" |
    "upload_video_note" |
    "upload_document";