// src/bot/types/model.d.ts
export type Chat =
  chatUser |
  chatGroup |
  chatSupergroup |
  chatChannel;

interface ChatBase {
    id: number;
    type: string;
    name: string;
    username?: string;

    smallPhoto?: File;
    bigPhoto?: File;
}

export interface chatUser extends ChatBase {
    type: "user";
    firstname: string;
    lastname: string | null;
    language?: string;
}

export interface chatGroup extends ChatBase {
    type: "group";
    title: string;
    allMembersAreAdmins?: boolean;
}

export interface chatSupergroup extends ChatBase {
    type: "supergroup";
    title: string;
    description?: string;
    inviteLink?: string;
}

export interface chatChannel extends ChatBase {
    type: "channel";
    title: string;
    description?: string;
    inviteLink?: string;
}

export type User = chatUser;

// Message types

export type Message =
  messageText |
  messageAudio |
  messageDocument |
  messagePhoto |
  messageSticker |
  messageVideo |
  messageVideoNote |
  messageVoice |
  messageContact |
  messageLocation |
  messageVenue |
  messageGame |
  messageUpdate;

interface MessageBase {
    id: number;
    type: string | undefined;
    date: Date;
    editDate?: Date;
    queued: boolean;
    edited?: boolean;

    chat: Chat;
    user?: chatUser;
    group?: chatGroup | chatSupergroup;

    from?: User;
    forward?: {
        date: Date;
        id?: string;
        chat?: Chat;
        from?: User;
    };
    reply?: Message;
}

interface messageTextBase extends MessageBase {
    type: "text";
    text: string;
    entities: MessageEntity[];

    mentions(): string[];

    hashtags(): string[];

    mentions(x: string): number;

    hashtags(x: string): number;
}

export type messageText = messageNonCommand | messageCommand;

export interface messageNonCommand extends messageTextBase {
    command: false;
}

export interface messageCommand extends messageTextBase {
    command: string;

    args(): string;

    args(N: number): string[];

    username: string | null;
    mine: boolean;
    exclusive: boolean;
}

export interface messageAudio extends MessageBase {
    type: "audio";
    duration: number;
    file: File;
    performer?: string;
    title?: string;
}
// ...
export interface File {
    id: string;
    size?: number;
    mime?: string;
    path?: string;
}